plugins/pikvm_client/requirements.txt
requests
pydantic
python-dotenv
websocket-client
grpcio
grpcio-tools
protobuf
pymongo
urllib3
pyotp
